requirements.txt
fastapi==0.110.0
importlib-metadata==4.6.4
Jinja2==3.1.3
numpy==1.26.4
pillow==10.2.0
torch==2.2.1
urllib3==2.2.1
uvicorn==0.28.1
torchvision
requests
ultralytics